+++ gpMgmt/bin/gppylib/db/catalog.py
@@ -10,8 +10,31 @@ import copy
 import dbconn
 from  gppylib import gplog
 from  pygresql import pg
+from gppylib.commands.base import Command, WorkerPool
 
-logger=gplog.get_default_logger()
+logger = gplog.get_default_logger()
+
+
+class RemoteQueryCommand(Command):
+    def __init__(self, qname, query, hostname, port, dbname=None):
+        self.qname = qname
+        self.query = query
+        self.hostname = hostname
+        self.port = port
+        self.dbname = dbname or os.environ.get('PGDATABASE', None) or 'template1'
+        self.res = None
+
+    def get_results(self):
+        return self.res
+
+    def run(self):
+        logger.debug('Executing query (%s:%s) for segment (%s:%s) on database (%s)' % (
+            self.qname, self.query, self.hostname, self.port, self.dbname))
+        with dbconn.connect(dbconn.DbURL(hostname=self.hostname, port=self.port, dbname=self.dbname),
+                            utility=True) as conn:
+            res = dbconn.execSQL(conn, self.query)
+            self.res = res.fetchall()
+
 
 class CatalogError(Exception): pass
